Bug 1483566 - Update Graphite2 to version 1.3.12. r=jfkthame
Differential Revision: https://phabricator.services.mozilla.com/D3426

+++ gfx/graphite2/src/GlyphFace.cpp
@@ -15,8 +15,8 @@
 
     You should also have received a copy of the GNU Lesser General Public
     License along with this library in the file named "LICENSE".
-    If not, write to the Free Software Foundation, 51 Franklin Street, 
-    Suite 500, Boston, MA 02110-1335, USA or visit their web page on the 
+    If not, write to the Free Software Foundation, 51 Franklin Street,
+    Suite 500, Boston, MA 02110-1335, USA or visit their web page on the
     internet at http://www.fsf.org/licenses/lgpl.html.
 
 Alternatively, the contents of this file may be used under the terms of the
@@ -33,16 +33,16 @@ int32 GlyphFace::getMetric(uint8 metric) const
 {
     switch (metrics(metric))
     {
-        case kgmetLsb       : return m_bbox.bl.x;
-        case kgmetRsb       : return m_advance.x - m_bbox.tr.x;
-        case kgmetBbTop     : return m_bbox.tr.y;
-        case kgmetBbBottom  : return m_bbox.bl.y;
-        case kgmetBbLeft    : return m_bbox.bl.x;
-        case kgmetBbRight   : return m_bbox.tr.x;
-        case kgmetBbHeight  : return m_bbox.tr.y - m_bbox.bl.y;
-        case kgmetBbWidth   : return m_bbox.tr.x - m_bbox.bl.x;
-        case kgmetAdvWidth  : return m_advance.x;
-        case kgmetAdvHeight : return m_advance.y;
+        case kgmetLsb       : return int32(m_bbox.bl.x);
+        case kgmetRsb       : return int32(m_advance.x - m_bbox.tr.x);
+        case kgmetBbTop     : return int32(m_bbox.tr.y);
+        case kgmetBbBottom  : return int32(m_bbox.bl.y);
+        case kgmetBbLeft    : return int32(m_bbox.bl.x);
+        case kgmetBbRight   : return int32(m_bbox.tr.x);
+        case kgmetBbHeight  : return int32(m_bbox.tr.y - m_bbox.bl.y);
+        case kgmetBbWidth   : return int32(m_bbox.tr.x - m_bbox.bl.x);
+        case kgmetAdvWidth  : return int32(m_advance.x);
+        case kgmetAdvHeight : return int32(m_advance.y);
         default : return 0;
     }
 }
